refactor: 댓글 삭제 및 좋아요 API와 훅 리팩토링
- deleteComment 및 likeComment 함수를 deleteCommentAPI와 likeCommentAPI로 이름 변경하여 API 호출 명확성 향상
- 댓글 삭제 및 좋아요 기능을 useDeleteComment 및 useLikeComment 훅으로 통합하여 코드 간소화
- PostsManagerPage에서 댓글 관련 상태 및 함수 제거
