+++ src/features/comment/like-comment/model/hooks.ts
@@ -1,4 +1,5 @@
 import { useCommentStore } from '../../../../entities/comment/model/store';
+import { likeCommentAPI } from '../api/api';
 
 export const useLikeComment = () => {
   const { comments, updateCommentInPost } = useCommentStore();
@@ -8,12 +9,7 @@ export const useLikeComment = () => {
       const currentComment = comments[postId]?.find((c) => c.id === id);
       if (!currentComment) return;
 
-      const response = await fetch(`/api/comments/${id}`, {
-        method: 'PATCH',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify({ likes: currentComment.likes + 1 }),
-      });
-      const data = await response.json();
+      const data = await likeCommentAPI(id, currentComment.likes);
       updateCommentInPost(postId, id, { ...data, likes: currentComment.likes + 1 });
     } catch (error) {
       console.error('댓글 좋아요 오류:', error);
